src/gateway-service/requirements.txt
Flask
requests
flask-cors
flask-jwt-extended
PyJWT